add map + refactor
