+++ src/Components/Map/Map.tsx
@@ -1,0 +1,59 @@
+import {useRef, useEffect} from 'react';
+import {Icon, Marker, layerGroup} from 'leaflet';
+import useMap from '../../hooks/useMap';
+import {City, Points, Point} from '../../types/points';
+import 'leaflet/dist/leaflet.css';
+
+type MapProps = {
+  city: City;
+  points: Points;
+  selectedPoint: Point | undefined;
+};
+
+const defaultCustomIcon = new Icon({
+  iconUrl: '../../../markup/img/pin.svg',
+  iconSize: [25, 40],
+  iconAnchor: [20, 40]
+});
+
+const currentCustomIcon = new Icon({
+  iconUrl: '../../../markup/img/pin-active.svg',
+  iconSize: [25, 40],
+  iconAnchor: [20, 40]
+});
+
+function Map(props: MapProps): JSX.Element {
+  const {city, points, selectedPoint} = props;
+
+  const mapRef = useRef(null);
+  const map = useMap(mapRef, city);
+
+  useEffect(() => {
+    if (map) {
+      map.setZoom(city.zoom);
+      const markerLayer = layerGroup().addTo(map);
+      points.forEach((point) => {
+        const marker = new Marker({
+          lat: point.lat,
+          lng: point.lng
+        });
+
+        marker
+          .setIcon(
+            selectedPoint !== undefined && point.id === selectedPoint.id
+              ? currentCustomIcon
+              : defaultCustomIcon
+          )
+          .addTo(markerLayer);
+      });
+
+      return () => {
+        map.removeLayer(markerLayer);
+      };
+    }
+  }, [map, points, selectedPoint, city.zoom]);
+
+  return <div style={{height: '500px'}} ref={mapRef}></div>;
+}
+
+export default Map;
