+++ src/Components/OfferPage/OfferPage.tsx
@@ -1,0 +1,282 @@
+import { useParams } from 'react-router-dom';
+import {Link} from 'react-router-dom';
+import { OfferDescription } from '../../types/offerDescription.ts';
+import { review } from '../../types/review.ts';
+import ReviewForm from '../../Components/ReviewForm/ReviewForm.tsx';
+
+function OfferPage({ offer, guestReview}: {offer:OfferDescription[];guestReview:review[]}):JSX.Element{
+  const { id } = useParams<{ id: string }>();
+  const filteredOffer:OfferDescription[] = offer.filter((of) =>(of.id === id));
+  return (
+
+    <div className="page">
+      <header className="header">
+        <div className="container">
+          <div className="header__wrapper">
+            <div className="header__left">
+              <a className="header__logo-link" href="main.html">
+                <Link to = "/">
+                  <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
+                </Link>
+              </a>
+            </div>
+            <nav className="header__nav">
+              <ul className="header__nav-list">
+                <li className="header__nav-item user">
+                  <a className="header__nav-link header__nav-link--profile" href="#">
+                    <div className="header__avatar-wrapper user__avatar-wrapper">
+                    </div>
+                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
+                    <span className="header__favorite-count">3</span>
+                  </a>
+                </li>
+                <li className="header__nav-item">
+                  <a className="header__nav-link" href="#">
+                    <span className="header__signout">Sign out</span>
+                  </a>
+                </li>
+              </ul>
+            </nav>
+          </div>
+        </div>
+      </header>
+
+      <main className="page__main page__main--offer">
+        <section className="offer">
+          <div className="offer__gallery-container container">
+            <div className="offer__gallery">
+              {filteredOffer[0].images.map((image) => (
+                <div key={image} className="offer__image-wrapper"> {}
+                  <img className="offer__image" src={image} alt="Фото студии"/>
+                </div>
+              ))}
+            </div>
+          </div>
+          <div className="offer__container container">
+            <div className="offer__wrapper">
+              {filteredOffer[0].isPremium ? (
+                <div className="offer__mark">
+                  <span>Premium</span>
+                </div>
+              ) : null}
+              <div className="offer__name-wrapper">
+                <h1 className="offer__name">
+                  {filteredOffer[0].offerName}
+                </h1>
+                <button className={filteredOffer[0].bookmark === 'In bookmarks' ? 'offer__bookmark-button offer__bookmark-button--active button' : 'offer__bookmark-button button'} type="button">
+                  <svg className="offer__bookmark-icon" width="31" height="33">
+                    <use xlinkHref="#icon-bookmark"></use>
+                  </svg>
+                  <span className="visually-hidden">{filteredOffer[0].bookmark}</span>
+                </button>
+              </div>
+              <div className="offer__rating rating">
+                <div className="offer__stars rating__stars">
+                  <span style={{width: filteredOffer[0].raitingStars}}></span>
+                  <span className="visually-hidden">Rating</span>
+                </div>
+                <span className="offer__rating-value rating__value">{filteredOffer[0].raiting}</span>
+              </div>
+              <ul className="offer__features">
+                {filteredOffer[0].features.map((feature) => (
+                  <li key={feature} className="offer__feature offer__feature--entire">
+                    {feature}
+                  </li>
+                ))}
+
+              </ul>
+              <div className="offer__price">
+                <b className="offer__price-value">&euro;{filteredOffer[0].price}</b>
+                <span className="offer__price-text">&nbsp;night</span>
+              </div>
+              <div className="offer__inside">
+                <h2 className="offer__inside-title">What&apos;s inside</h2>
+                <ul className="offer__inside-list">
+                  {filteredOffer[0].inside.map((ins) => (
+                    <li key={ins} className="offer__inside-item">
+                      {ins}
+                    </li>
+                  ))}
+                </ul>
+              </div>
+              <div className="offer__host">
+                <h2 className="offer__host-title">Meet the host</h2>
+                <div className="offer__host-user user">
+                  <div className="offer__avatar-wrapper offer__avatar-wrapper--pro user__avatar-wrapper">
+                    <img className="offer__avatar user__avatar" src={filteredOffer[0].hostAvatar} width="74" height="74" alt="Host avatar"/>
+                  </div>
+                  <span className="offer__user-name">
+                    {filteredOffer[0].hostName}
+                  </span>
+                  <span className="offer__user-status">
+                    {filteredOffer[0].hostStatus}
+                  </span>
+                </div>
+                <div className="offer__description">
+                  {filteredOffer[0].description.map((desc) => (
+                    <p key={desc} className="offer__text">
+                      {desc}
+                    </p>
+                  ))}
+
+
+                </div>
+              </div>
+              <section className="offer__reviews reviews">
+                <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">1</span></h2>
+                <ul className="reviews__list">
+                  {guestReview.map((rev) => (
+                    <li key={rev.name} className="reviews__item">
+                      <div className="reviews__user user">
+                        <div className="reviews__avatar-wrapper user__avatar-wrapper">
+                          <img
+                            className="reviews__avatar user__avatar"
+                            src={rev.img}
+                            width="54"
+                            height="54"
+                            alt="Reviews avatar"
+                          />
+                        </div>
+                        <span className="reviews__user-name">
+                          {rev.name}
+                        </span>
+                      </div>
+                      <div className="reviews__info">
+                        <div className="reviews__rating rating">
+                          <div className="reviews__stars rating__stars">
+                            <span style={{ width: rev.raitingStars }}></span>
+                            <span className="visually-hidden">Rating</span>
+                          </div>
+                        </div>
+                        <p className="reviews__text">
+                          {rev.description}
+                        </p>
+                        <time className="reviews__time" dateTime="2019-04-24">
+                          {rev.date}
+                        </time>
+                      </div>
+                    </li>
+                  ))}
+
+                </ul>
+                <ReviewForm
+                  onAnswer={() => {
+                    throw new Error('Function \'onAnswer\' isn\'t implemented.');
+                  }}
+                />
+              </section>
+
+            </div>
+          </div>
+          <section className="offer__map map"></section>
+        </section>
+        <div className="container">
+          <section className="near-places places">
+            <h2 className="near-places__title">Other places in the neighbourhood</h2>
+            <div className="near-places__list places__list">
+              <article className="near-places__card place-card">
+                <div className="near-places__image-wrapper place-card__image-wrapper">
+                  <a href="#">
+                    <img className="place-card__image" src="img/room.jpg" width="260" height="200" alt="Place image"/>
+                  </a>
+                </div>
+                <div className="place-card__info">
+                  <div className="place-card__price-wrapper">
+                    <div className="place-card__price">
+                      <b className="place-card__price-value">&euro;80</b>
+                      <span className="place-card__price-text">&#47;&nbsp;night</span>
+                    </div>
+                    <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+                      <svg className="place-card__bookmark-icon" width="18" height="19">
+                        <use xlinkHref="#icon-bookmark"></use>
+                      </svg>
+                      <span className="visually-hidden">In bookmarks</span>
+                    </button>
+                  </div>
+                  <div className="place-card__rating rating">
+                    <div className="place-card__stars rating__stars">
+                      <span style={{width: '80%'}}></span>
+                      <span className="visually-hidden">Rating</span>
+                    </div>
+                  </div>
+                  <h2 className="place-card__name">
+                    <a href="#">Wood and stone place</a>
+                  </h2>
+                  <p className="place-card__type">Room</p>
+                </div>
+              </article>
+
+              <article className="near-places__card place-card">
+                <div className="near-places__image-wrapper place-card__image-wrapper">
+                  <a href="#">
+                    <img className="place-card__image" src="img/apartment-02.jpg" width="260" height="200" alt="Place image"/>
+                  </a>
+                </div>
+                <div className="place-card__info">
+                  <div className="place-card__price-wrapper">
+                    <div className="place-card__price">
+                      <b className="place-card__price-value">&euro;132</b>
+                      <span className="place-card__price-text">&#47;&nbsp;night</span>
+                    </div>
+                    <button className="place-card__bookmark-button button" type="button">
+                      <svg className="place-card__bookmark-icon" width="18" height="19">
+                        <use xlinkHref="#icon-bookmark"></use>
+                      </svg>
+                      <span className="visually-hidden">To bookmarks</span>
+                    </button>
+                  </div>
+                  <div className="place-card__rating rating">
+                    <div className="place-card__stars rating__stars">
+                      <span style={{width: '80%'}}></span>
+                      <span className="visually-hidden">Rating</span>
+                    </div>
+                  </div>
+                  <h2 className="place-card__name">
+                    <a href="#">Canal View Prinsengracht</a>
+                  </h2>
+                  <p className="place-card__type">Apartment</p>
+                </div>
+              </article>
+
+              <article className="near-places__card place-card">
+                <div className="place-card__mark">
+                  <span>Premium</span>
+                </div>
+                <div className="near-places__image-wrapper place-card__image-wrapper">
+                  <a href="#">
+                    <img className="place-card__image" src="img/apartment-03.jpg" width="260" height="200" alt="Place image"/>
+                  </a>
+                </div>
+                <div className="place-card__info">
+                  <div className="place-card__price-wrapper">
+                    <div className="place-card__price">
+                      <b className="place-card__price-value">&euro;180</b>
+                      <span className="place-card__price-text">&#47;&nbsp;night</span>
+                    </div>
+                    <button className="place-card__bookmark-button button" type="button">
+                      <svg className="place-card__bookmark-icon" width="18" height="19">
+                        <use xlinkHref="#icon-bookmark"></use>
+                      </svg>
+                      <span className="visually-hidden">To bookmarks</span>
+                    </button>
+                  </div>
+                  <div className="place-card__rating rating">
+                    <div className="place-card__stars rating__stars">
+                      <span style={{width: '100%'}}></span>
+                      <span className="visually-hidden">Rating</span>
+                    </div>
+                  </div>
+                  <h2 className="place-card__name">
+                    <a href="#">Nice, cozy, warm big bed apartment</a>
+                  </h2>
+                  <p className="place-card__type">Apartment</p>
+                </div>
+              </article>
+            </div>
+          </section>
+        </div>
+      </main>
+    </div>
+  );
+}
+export default OfferPage;
