+++ src/mocks/city.ts
@@ -1,0 +1,8 @@
+import {City} from '../types/points';
+
+export const CITY: City = {
+  title: 'Амстердам',
+  lat: 52.374,
+  lng: 4.889969,
+  zoom: 12
+};
