+++ src/mocks/points.ts
@@ -1,0 +1,23 @@
+import { Points } from '../types/points.ts';
+export const POINTS:Points = [
+  {
+    id:1,
+    lat:52.3909553943508,
+    lng:4.85309666406198
+  },
+  {
+    id:2,
+    lat:52.3609553943508,
+    lng:4.85309666406198
+  },
+  {
+    id:3,
+    lat:52.3909553943508,
+    lng:4.929309666406198
+  },
+  {
+    id:4,
+    lat:52.3809553943508,
+    lng:4.939309666406198
+  }
+];
