+++ src/types/points.ts
@@ -1,0 +1,14 @@
+export type City = {
+  title: string;
+  lat: number;
+  lng: number;
+  zoom: number;
+};
+
+export type Point = {
+  id:string;
+  lat: number;
+  lng: number;
+};
+
+export type Points = Point[];
